requirements.txt
selenium
webdriver-manager
flask
pandas
psycopg2-binary
pydantic
pytest